#134 Fragment Event Consumer refactor.
