+++ actions/core/docs/asciidoc/dataobjects.adoc
@@ -12,7 +12,8 @@
 Sets <code>Action</code> configuration that is passed to Action.
 +++
 |[[doAction]]`@doAction`|`String`|+++
-Sets the name of the base <code>Action</code> that will be triggered while creating current <code>Action</code>. In not set (<code>null</code>), given action will have no base actions.
+Sets the <code>Action</code> (by its alias) that would be triggered by the returned action instance.
+ This is mainly used by <a href="https://github.com/Knotx/knotx-fragments/tree/master/actions/core#behaviours">behaviours</a>.
 +++
 |[[factory]]`@factory`|`String`|+++
 Sets <code>Action</code> factory name.
