+++ handler/consumer/html/src/main/resources/META-INF/services/io.knotx.fragments.handler.consumer.api.FragmentExecutionLogConsumerFactory
