+++ handler/consumer/json/src/test/java/io/knotx/fragments/handler/consumer/json/JsonFragmentsHandlerConsumerFactoryTest.java
@@ -13,22 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.knotx.fragments.handler.consumer;
+package io.knotx.fragments.handler.consumer.json;
 
-import io.knotx.fragments.handler.consumer.api.FragmentExecutionLogConsumer;
-import io.knotx.fragments.handler.consumer.api.FragmentExecutionLogConsumerFactory;
-import io.vertx.core.json.JsonObject;
+import static org.junit.jupiter.api.Assertions.*;
 
-public class TestConsumer implements FragmentExecutionLogConsumerFactory {
+// TODO fix #134
+class JsonFragmentsHandlerConsumerFactoryTest {
 
-  @Override
-  public String getName() {
-    return "testConsumer";
-  }
-
-  @Override
-  public FragmentExecutionLogConsumer create(JsonObject config) {
-    return (clientRequest, executionLogs) -> executionLogs
-        .forEach(log -> log.getFragment().setBody("testConsumer"));
-  }
 }
