+++ handler/core/src/test/resources/META-INF/services/io.knotx.fragments.handler.consumer.api.FragmentExecutionLogConsumerFactory
